+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,12 +1,24 @@
 # Release History
 
-## 0.5.8 (Unreleased)
+## 0.5.9 (Unreleased)
 
 ### Features Added
 
-- Add support for sample translation from one language to another
-
 ### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 0.5.8 (2025-11-13)
+
+### Features Added
+
+- Validate package names before adding to release plan
+- Allow forced creation when a release plan already exists
+- Add support for sample translation from one language to another
+- Updated CLI commands hierarchy to move all package related commands under `package` command group, all TypeSpec based commands under `tsp` command and similarly all release plan related commands under `release-plan`.
+- Added ChangelogContentUpdateTool, MetadataUpdateTool, and VersionUpdateTool
 
 ### Bugs Fixed